requirements.txt
python-pptx==0.6.23
google-generativeai==0.3.2
python-dotenv==1.0.1
streamlit==1.32.0 